requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-talent==2.6.0
python-dotenv==0.20.0
gunicorn==20.1.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
sklearn==0.0
scikit-survival==0.17.2
pyttsx3==2.90